refactoring: guest-app의 Question 관련 컴포넌트 폴더 구조 개선
# 컴포넌트 및 파일 이름 변경
* DeleteQuestionCardModal.js -> QuestionCardDeleteModal.js
* DeleteQuestionCardModal.js -> QuestionCardDeleteModal.js
* QuestionEditMenuDrawer.js -> QuestionCardEditMenuDrawer.js
* EditQuestoinCardMenuButton.js -> QuestionCardMenuEditButton
* UserInfoInput.js -> QuestionUserInfoInput.js

# 폴더명 변경
* QuestionArea -> QuestionInput

# 파일 위치 변경
* QuestionInputDrawer를 QuestionContainer 폴더로 이동

# etc
* 기타 불필요한 주석 제거
