+++ frontend/guest-app/src/components/QuestionCardEditMenuDrawer/QuestionCardDeleteModal.js
@@ -4,7 +4,7 @@ import Grid from "@material-ui/core/Grid";
 import Button from "@material-ui/core/Button";
 import CommonModal from "../CommonComponent/CommonModal/CommonModal.js";
 
-function DeleteQuestionCardModal(props) {
+function QuestionCardDeleteModal(props) {
 	const {isOpened, closeModal, onCancel, onDelete} = props;
 
 	return (
@@ -20,15 +20,15 @@ function DeleteQuestionCardModal(props) {
 	);
 }
 
-DeleteQuestionCardModal.propTypes = {
+QuestionCardDeleteModal.propTypes = {
 	isOpened: PropTypes.bool,
 	closeModal: PropTypes.func,
 	onCancel: PropTypes.func,
 	onDelete: PropTypes.func,
 };
 
-DeleteQuestionCardModal.defaultProps = {
+QuestionCardDeleteModal.defaultProps = {
 	isOpened: PropTypes.bool,
 };
 
-export default DeleteQuestionCardModal;
+export default QuestionCardDeleteModal;
