+++ frontend/guest-app/src/components/QuestionInput/QuestionContentInput.js
