+++ frontend/guest-app/src/components/QuestionInput/QuestionUserInfoInput.js
@@ -4,7 +4,7 @@ import useStringState from "../../hooks/useStringState.js";
 import UserAvatar from "../UserAvatar/UserAvatar.js";
 import QuestionUserNameInput from "./QuestionUserNameInput.js";
 
-function UserInfoInput(props) {
+function QuestionUserInfoInput(props) {
 	const {userNameRef, initialUserName = ""} = props;
 	const {state} = useStringState(initialUserName);
 
@@ -19,8 +19,8 @@ function UserInfoInput(props) {
 	);
 }
 
-UserInfoInput.propTypes = {
+QuestionUserInfoInput.propTypes = {
 	userNameRef: PropTypes.any,
 };
 
-export default UserInfoInput;
+export default QuestionUserInfoInput;
